+++ gegelatilib/include/instructions/addPrimitiveType.h
@@ -121,8 +121,6 @@ namespace Instructions {
     {
         this->operandTypes.push_back(typeid(T));
         this->operandTypes.push_back(typeid(T));
-        this->operandTypesNoConst.push_back(typeid(T));
-        this->operandTypesNoConst.push_back(typeid(T));
     }
 } // namespace Instructions