+++ gegelatilib/include/instructions/set.h
@@ -40,7 +40,6 @@
 
 #include <functional>
 
-#include "data/constant.h"
 #include "instructions/instruction.h"
 
 namespace Instructions {
@@ -103,15 +102,6 @@ namespace Instructions {
          * of operands needed to execute an Instruction in the Set.
          */
         unsigned int getMaxNbOperands() const;
-
-        /**
-         * \brief Get the maximum number of constants for Instruction currently
-         * contained in the Set.
-         *
-         * \return the unsigned int value corresponding to the largest number
-         * of constants needed to execute an Instruction in the Set.
-         */
-        unsigned int getMaxNbConstants() const;
     };
 } // namespace Instructions