+++ gegelatilib/include/program/program.h
@@ -254,19 +254,6 @@ namespace Program {
         Line& getLine(uint64_t index);
 
         /**
-         * \brief Return a vector corresponding to each constant in all the
-         * lines of the program
-         */
-        std::vector<double> getLineConstants() const;
-
-        /**
-         * \brief Set the current constants of all the lines in the program.
-         *
-         * \param[in] newConstants Vector of values to set to the constants
-         */
-        void setLineConstants(std::vector<double>& newConstants);
-
-        /**
          * \brief Checks whether a Line at the given index is an intron.
          *
          * \param[in] index The integer index of the checked Line within the