+++ gegelatilib/src/instructions/instruction.cpp
@@ -60,32 +60,7 @@ const std::vector<std::reference_wrapper<const std::type_info>>& Instruction::
 
 unsigned int Instructions::Instruction::getNbOperands() const
 {
-    return (unsigned int)this->operandTypesNoConst.size();
-}
-
-const std::vector<std::reference_wrapper<const std::type_info>>& Instruction::
-    getOperandTypesWithConst() const
-{
-    return this->operandTypes;
-}
-
-unsigned int Instructions::Instruction::getNbOperandsWithConst() const
-{
     return (unsigned int)this->operandTypes.size();
-}
-
-unsigned int Instructions::Instruction::getNbConstants() const
-{
-    unsigned int nbConstants = 0;
-
-    // Get the number of constant in the instruction
-    for (auto operandType : this->operandTypes) {
-        if (operandType.get() == typeid(Data::Constant)) {
-            nbConstants++;
-        }
-    }
-
-    return nbConstants;
 }
 
 bool Instruction::checkOperandTypes(
@@ -147,7 +122,7 @@ std::string Instruction::getPrintablePrimitiveOperandType(
         type = cm[2].str();
     }
     if (type == DEMANGLE_TYPEID_NAME(typeid(Data::Constant).name())) {
-        type = "int32_t";
+        type = "double";
     }
     // Default case
     return type;