+++ gegelatilib/src/program/line.cpp
@@ -129,31 +129,3 @@ bool Program::Line::operator!=(const Line& other) const
 {
     return !(this->operator==(other));
 }
-
-uint64_t Program::Line::getNbConstants() const
-{
-    return this->nbConstants;
-}
-
-void Program::Line::setNbConstants(uint64_t newNbConstants)
-{
-    this->nbConstants = newNbConstants;
-}
-
-const Data::ConstantHandler& Program::Line::cGetConstantHandler() const
-{
-    return this->constants;
-}
-
-Data::ConstantHandler& Program::Line::getConstantHandler()
-{
-    return this->constants;
-}
-
-const Data::Constant Program::Line::getConstantAt(size_t index) const
-{
-    std::shared_ptr<const Data::Constant> value =
-        this->constants.getDataAt(typeid(Data::Constant), index)
-            .getSharedPointer<const Data::Constant>();
-    return *value;
-}